+++ visualization/app/codeCharta/state/settingsService/settings.service.ts
@@ -167,7 +167,8 @@ export class SettingsService implements FileStateServiceSubscriber {
 				whiteColorBuildings: false,
 				mapColors: mapColors,
 				isPresentationMode: false,
-				showOnlyBuildingsWithEdges: false
+				showOnlyBuildingsWithEdges: false,
+				resetCameraIfNewFileIsLoaded: true
 			},
 			treeMapSettings: {
 				mapSize: 500