+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ### Added
 
+- Checkbox in global Settings for disabling camera reset, when new map is loaded #685
+- Pipe support for SourceCodeParser #716
+- Pipe support for SCMLogParser #717
+
 ### Changed
 
 ### Removed